# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TransMóvil - Transport & Local Services Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers/available:
    get:
      operationId: getAvailableDrivers
      tags: [drivers]
      summary: Get available drivers (for customers)
      parameters:
        - name: serviceTypeId
          in: query
          schema:
            type: integer
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Available drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DriverPublic"
  /drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get driver public profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverPublic"
        "404":
          description: Not found
  /drivers/me:
    get:
      operationId: getMyDriverProfile
      tags: [drivers]
      summary: Get current driver profile
      responses:
        "200":
          description: Driver profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfile"
        "404":
          description: Not a driver
    post:
      operationId: registerDriver
      tags: [drivers]
      summary: Register as a driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverRegistration"
      responses:
        "201":
          description: Driver registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfile"
    patch:
      operationId: updateDriverProfile
      tags: [drivers]
      summary: Update driver profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfile"
  /drivers/me/availability:
    patch:
      operationId: updateDriverAvailability
      tags: [drivers]
      summary: Toggle driver availability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverAvailabilityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfile"
  /drivers/me/location:
    patch:
      operationId: updateDriverLocation
      tags: [drivers]
      summary: Update driver's current location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: Location updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfile"
  /drivers/me/earnings:
    get:
      operationId: getMyEarnings
      tags: [drivers]
      summary: Get driver earnings summary
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, all]
            default: week
      responses:
        "200":
          description: Earnings summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarningsSummary"
  /drivers/me/trips:
    get:
      operationId: getMyDriverTrips
      tags: [drivers]
      summary: Get driver's service history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Trip list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"

  # ─── VEHICLES ─────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get user notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /admin/drivers/{id}:
    patch:
      operationId: updateAdminDriver
      tags: [admin]
      summary: Approve / suspend / update driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDriverUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfile"
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, role, firstName, lastName, email, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        role:
          type: string
          enum: [customer, driver, admin]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    DriverPublic:
      type: object
      required: [id, firstName, lastName, rating, totalTrips, isAvailable, serviceTypeId, lat, lng]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: number
        totalTrips:
          type: integer
        isAvailable:
          type: boolean
        serviceTypeId:
          type: integer
        serviceTypeName:
          type: string
        vehicleMake:
          type: ["string", "null"]
        vehicleModel:
          type: ["string", "null"]
        vehicleColor:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
        vehiclePhotoUrl:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
    DriverProfile:
      type: object
      required: [id, userId, status, isAvailable, rating, totalTrips, serviceTypeId, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, approved, suspended]
        isAvailable:
          type: boolean
        rating:
          type: number
        totalTrips:
          type: integer
        totalEarnings:
          type: number
        serviceTypeId:
          type: integer
        bio:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        serviceType:
          $ref: "#/components/schemas/ServiceType"
    DriverRegistration:
      type: object
      required: [serviceTypeId]
      properties:
        serviceTypeId:
          type: integer
        bio:
          type: string
    DriverProfileUpdate:
      type: object
      properties:
        serviceTypeId:
          type: integer
        bio:
          type: string
    DriverAvailabilityUpdate:
      type: object
      required: [isAvailable]
      properties:
        isAvailable:
          type: boolean
        lat:
          type: number
        lng:
          type: number
    LocationUpdate:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
    EarningsSummary:
      type: object
      required: [period, totalEarnings, platformCommission, driverEarnings, totalTrips]
      properties:
        period:
          type: string
        totalEarnings:
          type: number
        platformCommission:
          type: number
        driverEarnings:
          type: number
        totalTrips:
          type: integer
        recentPayments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
    Vehicle:
      type: object
      required: [id, driverId, make, model, year, plate, color, serviceTypeId]
      properties:
        id:
          type: integer
        driverId:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        plate:
          type: string
        color:
          type: string
        serviceTypeId:
          type: integer
        photoUrl:
          type: ["string", "null"]
        capacity:
          type: integer
        createdAt:
          type: string
...
    Trip:
      type: object
      required: [id, customerId, status, pickupAddress, pickupLat, pickupLng, destinationAddress, destinationLat, destinationLng, serviceTypeId, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        driverId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [searching, accepted, arriving, in_progress, completed, cancelled]
        pickupAddress:
          type: string
        pickupLat:
          type: number
        pickupLng:
          type: number
        destinationAddress:
          type: string
        destinationLat:
          type: number
        destinationLng:
          type: number
        serviceTypeId:
          type: integer
        estimatedPrice:
          type: ["number", "null"]
        finalPrice:
          type: ["number", "null"]
        paymentMethod:
          type: ["string", "null"]
        cancelReason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        acceptedAt:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        customer:
          $ref: "#/components/schemas/UserProfile"
        driver:
          $ref: "#/components/schemas/DriverPublic"
        serviceType:
          $ref: "#/components/schemas/ServiceType"
...
    ServiceType:
      type: object
      required: [id, name, icon, description, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        basePrice:
          type: number
        pricePerKm:
          type: number
        isActive:
          type: boolean
    Notification:
      type: object
      required: [id, userId, type, title, body, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [trip_request, trip_accepted, driver_arriving, trip_started, trip_completed, trip_cancelled, system]
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        tripId:
          type: ["integer", "null"]
        createdAt:
          type: string
    Payment:
      type: object
      required: [id, tripId, driverId, amount, driverAmount, platformAmount, method, createdAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        driverId:
          type: integer
        amount:
          type: number
        driverAmount:
          type: number
        platformAmount:
          type: number
        method:
          type: string
        status:
          type: string
        createdAt:
          type: string
...
    AdminDriverUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, suspended]
        notes:
          type: string
...