# lib/api-zod/src/generated/types/driverAvailabilityUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TransMóvil - Transport & Local Services Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DriverAvailabilityUpdate {
  isAvailable: boolean;
  lat?: number;
  lng?: number;
}