# lib/api-zod/src/generated/types/driverProfileUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TransMóvil - Transport & Local Services Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DriverProfileUpdate {
  serviceTypeId?: number;
  bio?: string;
}